requirements.txt
gradio==4.22.0
python-dotenv==1.0.0
